+++ scheduler/main.c
@@ -696,7 +696,7 @@ main(int  argc,				/* I - Number of command-line args */
   current_time  = time(NULL);
   event_time    = current_time;
   expire_time   = current_time;
-  local_timeout = current_time + 60;
+  local_timeout = 0;
   fds           = 1;
   report_time   = 0;
   senddoc_time  = current_time;
@@ -958,7 +958,10 @@ main(int  argc,				/* I - Number of command-line args */
     */
 
     if (current_time >= local_timeout)
+    {
       cupsdDeleteTemporaryPrinters(0);
+      local_timeout = 0;
+    }
 
 #ifndef HAVE_AUTHORIZATION_H
    /*
@@ -1732,11 +1735,11 @@ select_timeout(int fds)			/* I - Number of descriptors returned */
 
   for (printer = (cupsd_printer_t *)cupsArrayFirst(Printers); printer; printer = (cupsd_printer_t *)cupsArrayNext(Printers))
   {
-    if (printer->temporary && !printer->job && local_timeout > (printer->state_time + 60))
+    if (printer->temporary && !printer->job && (!local_timeout || local_timeout > (printer->state_time + 60)))
       local_timeout = printer->state_time + 60;
   }
 
-  if (timeout > local_timeout)
+  if (timeout > local_timeout && local_timeout)
   {
     timeout = local_timeout;
     why     = "delete stale local printers";